\name{crc32c-package}
\alias{crc32c-package}
\docType{package}
\title{\packageTitle{crc32c}}
\description{\packageDescription{crc32c}}
\section{Package Content}{\packageIndices{crc32c}}
\author{The CRC32C Authors for the \sQuote{crc32c} library;
  Dirk Eddelbuettel for the package.}
\section{Maintainer}{Dirk Eddelbuettel}
\keyword{package}
